Fix groff build with long interpeter line (#21549)
This PR fixes the case where groff fails to build if the spack install
path is really long. There are a couple of perl scripts that get built,
and used, during the build phase that will fail when the perl
interpreter line is too long. Filtering the lines will not work because
the files don not exist after the configure phase and patching after the
build phase is too late.  This PR runs the scripts explicitly with the
spack perl via the $(PERL) variable in the call to the script.