Revert "`url_exists` related improvements (#34095)" (#34144)
This reverts commit 1279979fd0215c9442277c4c551508a6f7e2ff5e.

The problem is that Bitbucket's API forwards download requests to an S3 bucket using a temporary URL. This URL includes a signature for the request, which embeds the HTTP verb. That means only GET requests are allowed, and HEAD requests would fail verification, leading to 403 erros. The same is observed when using `curl -LI ...`