Concretization Cache: Refactor + manifest testing (#49589)
#48198 Introduces the concretization cache, and associated manifest. However,
it only added testing for the cache itself, not the manifest or cleanup procedures.

This PR:

- [x] Adds tests for the manifest and cleanup
- [x] Refactors the manifest to update correctly
- [x] Refactors the cache to store base32 hashes like the rest of Spack's hashes
- [x] gzips the cache entries for a smaller memory footprint

`FileCache` class was refactored into an abstract base `Cache` class which provides
the general cache read/write transaction interface, and a `FileCache` and
`DirectoryCache` child classes, `FileCache` which should be functionally identical to
the `FileCache` class pre refactor, deals with caches at the individual file level, and
`DirectoryCache`, which is mean to provide a directory level or "bucket" caching level.