Checks for databases and mirror indexes (#50859)
Several bugs for the logic regarding cache indexes and contents.json have been exposed and tried to fix in this PR. It is currently hot-fix without added tests, but relevant to #47615 .

If the index file gets garbage collected/removed, but still remains in the contents.json/objects memory, spack will try to open the file and fail. This PR adds a write transaction, and we construct new data if necessary before reading.

If the remote hash gets out of sync the code can get into a state where it thinks it should update the index, but then recomputes the same file name for the new index file. In this case it currently garbage collects itself. Added logic to skip if the new index file computed already exists.

Database reads don't use file-locks, but also have no check for existence of the database file pre-read attempt. Add an existence check for cleaner debugging.

I think these are all valid bugs and fixes. It indicates that binary_distribution.py and the tests for it need a lot more attention/refactoring post 1.0 no matter what. Items 2 and 3 should not be required for #47615, but were helpful during debugging iterations. They are optional, but I think they should be added. Especially item 2.


---------

Co-authored-by: Greg Becker <becker33@llnl.gov>