Spec._format_dependencies: special case concrete specs to avoid stalling  (#50892)
Spec._format_dependencies recurse for transitive deps, and follows every path. 
The time to run this function thus increases exponentially for concrete specs.

This doesn't solve the issue for abstract specs with a lot of transitive edges, 
and a deep DAG structure, but these cases are rare and encountered only 
when a concrete spec is being built by the concretizer.

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>