Stop sharing misc_cache among Spack instances (#50595)
There are several caches Spack maintains within `config:misc_cache`
that are sensitive to changes in Spack; that location defaults to
`~/.spack/cache`, so when two different Spack instances are used
on the same system, they can conflict. Prior workarounds have
included:

* Running `spack clean -m` (often a temporary solution as the problem
  tends to recur when swapping back)
* Setting `SPACK_USER_CACHE_PATH` (this also moves some other caches
  which are fine to share, but relocates `config:misc_cache` as well)

This PR moves the default for `config:misc_cache` to a subdirectory
within `~/.spack` that is unique for for the instance of spack that
maintains it, based on a hash of Spack's prefix. With that, Spack
instances should be sufficiently isolated without having to set
`SPACK_USER_CACHE_PATH` (or `config:misc_cache`).

---------

Co-authored-by: Patrick Gartung <gartung@fnal.gov>
Co-authored-by: Peter Scheibel <scheibel1@llnl.gov>