config: revert default `misc_cache` to `~/.spack/cache` (#50920)
The `$spack_instance_id` variable was more breaking than anticipated, in particular for CI environments where the same Spack commit is cloned to unique paths.

This commit leaves the variable `$spack_instance_id` available for use in config, but reverts the changes to default config. This means `misc_cache` remains in `~/.spack/cache`.

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>