Spec.installed_upstream fix (#50938)
`Database.query_by_spec_hash` returns `upstream, record`, and
claims in its docstring that `upstream = True` if the spec is installed
upstream. However, the record only needs to be *present* upstream
for this to be true. `Spec.installed_upstream` was going by the docstring
(which meant it could be wrong in circumstances where the upstream
had a record, but the spec wasn't actually installed there).

I think it's useful for `query_by_spec_hash` to preserve its behavior, so I

* Updated the docstring
* Made `Spec.installed_upstream` check the record

I also added a test. More than just the behavior here, I didn't see an explicit
test for the case where a spec is added to the DB with no prefix - something
that can now happen fairly frequently with the compiler wrappers when installing
from a binary cache (which is also how a user found this problem).