solver: fix issues with identical externals (modulo compiler) (#50956)
fixes #50955

Externals are coded with a pair of "trigger conditions"
and "imposed conditions".

In the case of the issue being solved, the trigger conditions
were the same for both mpich:
```
condition_requirement(22,"node","mpich").
condition_requirement(22,"node_version_satisfies","mpich","=4.3.0").
```
but the imposed conditions were conflicting.

To fix the issue, relax the set of rules and allow clingo
to not impose an external condition.

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>