Set the default for builds to be p1j16 (#50973)
Set the default for builds to be `--concurrent-packages 1` 
and `--jobs 16` (i.e., `-p1 -j16`).

Packages will install one at a time with 16 build jobs unless otherwise set
with `--concurrent-packages/--jobs` on the command line or in config with
* `config:concurrent_packages` or
* `config:build_jobs`

The main reasons for doing this are:
1. Without `gmake` job server support, `-p4 -j4` is slower than `-p1 -j16` for
   large package builds like `llvm` and `trilinos`; and
2. Currently, Spack does not ensure that parallel build output is synced line-by-line
   or task-by-task. Without this, output can appear messy by default.

When we've added (1) and (2), we'll make concurrent builds the default again.