docs: add and apply rst formatter (#51144)
Apply a formatter on rst files from our docs and enable it in GitHub Actions.

The formatter ensures that sentences are on a single line. End of sentence is determined using a *heuristic*, this is not adding any advanced natural language processing tokenizers.

The motivation for the formatter is two-fold:

1.  Diffs will be easier to review when sentences are modified. These types of changes are pretty common. Adding or removing a word currently triggers a cascade of changes to respect line length limits, ultimately hiding the actual change. After this, you see the world level diff highlighted in the GitHub UI (locally: `git diff --word-diff`), and no more scanning whether other parts of the sentence were modified on other lines is needed.
2. It catches a common issue where a missing blank line before a list results in the list being parsed as text in the paragraph:
   ```
   Example:
   1. First item
   2. Second item
   ```
   This example fails the check, because it's parsed as a paragraph by docutils and then reformatted as one-sentence-per-line:
   ```
   Example: 1.
   First item 2.
   Second item
   ```
   Users can then fix it by adding a newline between `Example:` and `1. First item`.

The formatter was tested and produces the same normalized HTML for the docs before and on this commit using the `tidy -i -wrap 0 -m *.html */*.html` command. 

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>