docs: disable values in autodoc (#51272)
Do not document default values, cause they are documented as values as
Sphinx evaluates them, instead of the Python expressions they are
defined as.

An example of where this is currently really bad is
`spack.config.SECTION_SCHEMAS`, which is an enormous, deeply nested data
structure.

Another example that was potentially a security issue is where we
re-exported `os.environ` in the package API, and ended up "documenting"
the environment variables inside the read the docs build, which could
leak secrets.

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>