schema: fix issues with regexes (#51295)
In various cases, people tried to express "valid identifier" as
`\w[\w-]*` but glanced over the fact that it needs `^...$`. So, most of
the `patternProperties` we've used were simply wrong.

Then, a second issue is that some of these patterns were actually wrong
for their use case, such as matching specs in projections, where
`\w[\w-]+` was used, which would not match `^zlib` when wrapping it in
`^...$`. Our docs however advertise that in examples.

The cleanest fix seems to be to just use `additionalProperties` in general,
and fix a few regexes in `patternProperties` that are needed.