installer.py: start FIFO jobserver (#50708)
If Spack itself is not running under a POSIX jobserver and gmake and ninja from the DAG support it, start a job server using a named pipe before running the installer.

This ensures that gmake and ninja subprocesses of Spack will limit parallelism. It does not limit parallelism of Spack's own installer yet, it only implements the server side of the jobserver.

Signed-off-by: Kathleen Shea <shea9@llnl.gov>