Enable concretization caching (redux) (#50905)
Third try to enable concretization caching in `develop`. Past attempts were reverted due to cache
management issues and race conditions.

See #48198 for the introduction of the concretization cache, manifest, etc. That PR only
added testing for the cache itself, not the manifest or cleanup procedures.

This PR:

- [x] Adds tests for cleanup
- [x] Refactors the cache to store base32 hashes like the rest of Spack's hashes
- [x] gzips cache entries for a smaller memory footprint
- [x] Removes redundant stat calls 
- [x] Less aggressive/error prone cleanup phase
- [x] Adds an entry limit to the cache, removes the size limit
- [x] Rewrite pruning to use LRU instead of FIFO
- [x] Canonicalizes and strips `clingo` input before hashing, for greater hit rate

This PR simplifies the locking around the concretization cache. Entries are stored by
the base32 hash of their inputs, and there is a single lock file (as we do in many places in
spack) for all such entries. We use byte range locks to manage all the entries.

Cleanup has changed to be less expensive overall. There is no more size limit; there is only
an entry limit, and specs in the cache are compressed. Specs are ~35k, and the initial limit
is 1000 cache entries, or ~35 MB.  This limit is customizable in configuration. To check if the
cache needs cleanup, we need only list files in the cache directory. To clean up, we stat the files
and clean up the least recently used concretizations.  We remove half the cache at a time so that
we do not pay for cleanup on every concretization when we're close to the cache limit.

The concretization cache is stored in the `misc_cache` by default, so `spack clean -m` is all
that is needed to get rid of it.  The location of the concretization cache can be specified
independently (with a `url` config option) but we will document and expand on that use later
as we integrate it in CI.

---------

Signed-off-by: John Parent <john.parent@kitware.com>
Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>
Co-authored-by: Todd Gamblin <tgamblin@llnl.gov>