Configuration scopes: fix severe ordering issue when `-C` and `-e` are used together (#51461)
Fixes #51459.

#50853 attempted to make configuration (envs and config scopes) specified on the CLI
merge in CLI order, but did not quite succeed.  Environments, when loaded, would still
appear out of order in the scope stack. This fixes the issues and establishes, after much
discussion, the ordering we want for configuration scopes specified on the CLI:

1. Ordering of scopes will be enforced *within* the same option on the CLI, e.g. bar is higher precedence in `-C foo -C bar`;
2. Ordering *across* options is counterintuitive, and the `-e` option (or the active environment via `SPACK_ENV`) should always be lower priority than `-C` or `-c` options; and
4. `-C` options have lower priority than `-c`.

Environments are now added using `ConfigScopePriority.ENVIRONMENT` in all cases.
Since they are the only entity using that scope priority, it's safe to activate and deactivate them.

Documentation has been updated to reflect that the CUSTOM scopes have higher priority
than ENVIRONMENT.

---------

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>