solver: set preferences from input specs using `%%` (#51383)
This PR allows using `%%` to set propagated preferences on input
specs, and is particularly useful when dealing with `unify: false`
environments.

An example of an environment that can be concretized with this PR is:

```yaml
spack:
  definitions:
  - compilers: ["%%llvm_gfortran", "%%c,cxx,fortran=gcc"]
  specs:
  - matrix:
    - [mpileaks, hdf5, trilinos]
    - [$compilers]
  toolchains:
    llvm_gfortran:
    - spec: "%c=llvm"
      when: "%c"
    - spec: "%cxx=llvm"
      when: "%cxx"
    - spec: "%fortran=gcc"
      when: "%fortran"
  packages:
    callpath:
      require:
      - "%c=gcc"
  concretizer:
    unify: false
```

`%%` is now a *propagated* preference. i.e., it forces everything in a node's
sub-DAG to use the preference unless it is impossible.

To allow for exceptions using `%`, `%%`, and spec syntax while still allowing
global preferences in `packages.yaml`,  the penalty taken for not fulfilling a `%%`
preference is higher than that of a preference in `packages.yaml`. In this way we
can propagate compilers to the `link,run` closure of a node, and set global
preferences for other nodes.

The performance penalty for using `%%` is not insignificant, but it is also not
present when *not* using `%%`. This has essentially zero cost for solves that do
not use the new functionality.

---------

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>