concretization: improve error messages for requirements (#45800)
This PR adds a number of tests that exercise error messages and
adjusts weights on error messages in the solver to improve the quality
of messages of the tests.

The tests added are all designed to create a concretization attempt that will fail
because the problem is over-constrained. Some tests add constraints to package
definitions; some add requirements to configuration. The tests verify that a good
error message is achieved for each test case.

Specific improvements:

* The concretizer has the freedom to choose a wrong version. The error weights
  for this type of message was low, so it could short-circuit a cause chain by just
  assigning a wrong version. The version error mechanism is now split in two: 

  1. Arbitrary version selections that lead to version errors are penalized heavily.
  2. Version errors that result from other constraints are not.

* When this PR was first created, the error messages for requirements were
  particularly bad. They have been broadly improved by heavily encouraging
  requirement groups to apply at least one of their constraints, which appears
  to properly hook into the causal chaining logic in `error_messages.lp`.

---

Signed-off-by: Peter Scheibel <scheibel1@llnl.gov>