`config`: use `prefer_modify: true` to adjust edit preferences (#51510)
* `config`: use `prefer_modify: true` to adjust edit preferences

#51162 allows users to specify configuration scopes as configuration, but it also
changes the `spack` scope to be the default edit scope. This can be extremely confusing
to users, as it causes `spack compiler find`, `spack external find`, `spack config add`,
etc. to be written into the spack instance by default instead of into `~/.spack`.

`~/.spack` is a better choice if it's configured, as we do not know if the user
actually owns the spack instance. It's also what people are used to, and it's going
to cause a lot of consternation if we switch this default up on people.

This PR introduces a new `prefer_modify: true` option on `include:` that allows the
including scope to hand off edit precedence to an included scope. The including scope
is still modifiable if specified explicitly by name (e.g., `--scope spack`) but if it
is selected by default, we'll edit its `prefer_modify` scope instead.

- [x] Add `prefer_modify:` option to `include:`
- [x] Make the default `spack` scope specify the user scope as `prefer_modify:`.
- [x] Add documentation and update `include:` docs with `versionadded` directives.
- [x] make the `user` scope `prefer_modify:` in `$spack/etc/spack/include.yaml`

---------

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>