Docs: Add package review guide and roles and responsibilities (#51154)
This PR integrates content from the `Spack Package Merge Criteria`
document into the `Contributing` section of the Spack documentation.

It separates the two major sections of the original document,
Roles and Responsibilities and Merge Criteria, into separate sections
within the Spack documentation. The merge criteria has been
restructured into more of a review guide, with additional contents
and lots of links to relevant areas of the Spack documentation.

---------

Signed-off-by: tldahlgren <dahlgren1@llnl.gov>
Co-authored-by: Harmen Stoppels <harmenstoppels@gmail.com>