bootstrap: update the versions of a few CMake dependencies (#51522)
* bootstrap: update the versions of a few CMake dependencies

The current JSON prototype contains a few versions that have been
deprecated, or removed, in the builtin repository.

Before Spack v1.0 it would have been fine to update the JSON file.
Since the core tool and the packages were tied together, we knew
which versions we could expect from a given commit of Spack.

Since now the builtin repository is detached, and we don't know
at which commit it's used, it would be better not to assume that
a certain version of a dependency is available.

Therefore, try to use, for the nodes that get frequent deprecations,
the "best" version according to the current configuration.

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>

* Filter cmake and dependencies by valid range

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>

* bootstrap: deal with patches, if we update versions

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>

* Move _inject_patches_variant to deal with circular imports

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>

---------

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>
Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>