new_installer.py: implement spack install -pN
In the new installer, `-j` is a global number of jobs shared by
concurrent package builds. The `-p` flag was so far not implemented, and
parallelism was effectively only limited by `-j`.

Sometimes though, it's useful to limit the number of concurrent package
builds when disk space in the stage dir is limited, or concurrent
fetching is undesirable.

With this change, package concurrency can be limited with `-pN`, while
keeping the CPU load constant (in an ideal world where the POSIX
jobserver is respected and the limited number of packages have
sufficient parallelism themselves).

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>