env: drop default venv search path from `PYTHONPATH` when activating Spack env  (#51743)
Environment views that contain Python packages are by design Python virtual environments. The Python interpreter knows how to locate Python packages inside the virtual environment without the PYTHONPATH environment variable as long as they are installed in the standard `site-packages` dir. Before this commit, Spack set PYTHONPATH to that dir. That's not only redundant, but also undesirable, because it makes other Python interpreters consider this virtual environment too now, whereas it's meant to be self-contained and isolated.

Therefore, prevent that the virtual environment's default search path is prepended to PYTHONPATH when activating an environment.

Signed-off-by: HUARD Ivan 284669 <ih284669@is248415.intra.cea.fr>