concretizer: parallel concretization on darwin/win32/linux-python3.14 (#51680)
* concretize.py: parallelism on darwin/macos/linux-python3.14

Now that we serialize environments and specs using pickling, we can
safely enable parallel concretization on Darwin and macOS and reenable
on Linux with Python 3.14.

The only change now is to actually serialize the active environment when
spawning or forkservering a subprocess.

Also, add a monkeypatch autouse fixture to disable multiprocessing
during tests except for one "smoke test", to avoid oversubscription,
since we use pytest-xdist already.

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>

* remove custom attrs repo_token and store_token

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>

---------

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>