new_installer.py: support forkserver
This adds support for Python 3.14 in the new installer. The issue is
that the classic self-pipe trick does not work for `forkserver`, because
there the forkserver process receives the signal that a child has
finished, not the parent process.

What's more, it avoids the O(n) search for finished subprocesses,
whether forkserver or not. Finally it's slightly more non-unix
compatible (not that Windows is supported now).

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>